% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/estimateit.R
\name{estimateit}
\alias{estimateit}
\title{Estimate marginal effects for binary exposure and outcome}
\usage{
estimateit(weightitobj, outcome, data)
}
\arguments{
\item{weightitobj}{A Weightit object}

\item{outcome}{A binary outcome variable coded 0/1}

\item{df}{A data frame containing the variables}
}
\value{
Returns a Summary tibble, and the individual effects
}
\description{
Estimate marginal effects for binary exposure and outcome
}
\examples{
library(cobalt)
library(WeightIt)

W1 <- weightit(treat ~ age + educ + married +
                re74, data = lalonde,
              method = "ps", estimand = "ATT")
summary(W1)
bal.tab(W1)
E1 <- estimateit(W1, nodegree, lalonde)
E1
}
