% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/estimateit.R
\name{estimateit}
\alias{estimateit}
\title{Estimate marginal effects for binary exposure and outcome}
\usage{
estimateit(weightitobj, outcome, data)
}
\arguments{
\item{weightitobj}{A Weightit object}

\item{outcome}{A binary outcome variable}

\item{data}{A data frame containing the outcome}
}
\value{
Returns a summary table as a tibble, the model, and the individual effects
   (X0 = "Control", X1="Treat", D=their difference, RR= their relative "risk",
   OR=their odds ratio)
}
\description{
Estimate marginal effects for binary exposure and outcome
}
\examples{
library(cobalt)
library(WeightIt)

W1 <- weightit(treat ~ age + educ + married +
                re74, data = lalonde,
              method = "ps", estimand = "ATT")
summary(W1)
bal.tab(W1)
E1 <- estimateit(W1, nodegree, lalonde)
E1
}
